refactor: 文件/目录命名规范化为 snake_case (PEP 8)
- 重命名 9 个根目录文件、4 个目录、4 个 functions 内文件

- 更新约 30 处导入语句

- 修复 filter_type 返回 None 导致的迭代错误
