+++ operators/attribute_ops.py
@@ -7,8 +7,8 @@ Mesh Attribute 标记 Operators
 """
 
 import bpy
-from ..Const import *
-from ..Functions.CommonFunctions import *
+from ..const import *
+from ..functions.common_functions import *
 
 
 class HST_OT_MarkTintObject(bpy.types.Operator):
