+++ operators/bevel_ops.py
@@ -7,9 +7,9 @@ Bevel 操作 Operators
 """
 
 import bpy
-from ..Const import *
-from ..Functions.HSTFunctions import *
-from ..Functions.CommonFunctions import *
+from ..const import *
+from ..functions.hst_functions import *
+from ..functions.common_functions import *
 
 
 class HST_OT_BevelTransferNormal(bpy.types.Operator):
