+++ operators/collection_ops.py
@@ -7,8 +7,8 @@ Collection 管理 Operators
 """
 
 import bpy
-from ..Const import *
-from ..Functions.CommonFunctions import *
+from ..const import *
+from ..functions.common_functions import *
 
 
 class HST_OT_MarkDecalCollection(bpy.types.Operator):
