+++ operators/collision_ops.py
@@ -7,8 +7,8 @@
 """
 
 import bpy
-from ..Const import *
-from ..Functions.CommonFunctions import *
+from ..const import *
+from ..functions.common_functions import *
 
 
 class HST_OT_SetUECollision(bpy.types.Operator):
