+++ operators/decal_ops.py
@@ -7,8 +7,8 @@ Decal Collection 操作 Operators
 """
 
 import bpy
-from ..Const import *
-from ..Functions.CommonFunctions import *
+from ..const import *
+from ..functions.common_functions import *
 
 
 class HST_OT_ActiveCollection(bpy.types.Operator):
