+++ rigging.py
